backend/pyproject.toml
[project]
name = "calendar-assistant-backend"
version = "0.1.0"
description = "Calendar Assistant API (FastAPI)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.100.0",
    "PyJWT>=2.8.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-ollama>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "dateparser", 
    "geopy>=2.4.1",
    "timezonefinder>=8.2.1"
]

[project.optional-dependencies]
dev = ["httpx"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
